step3 refactor : car의 패키지를 domain으로 변경
