+++ src/main/kotlin/step3/domain/CarFactory.kt
@@ -1,7 +1,5 @@
 package step3.domain
 
-import step3.model.Car
-
 fun interface CarFactory {
     fun create(): Car
 }
