+++ src/main/kotlin/step3/domain/RandomEngineCarFactory.kt
@@ -1,7 +1,5 @@
 package step3.domain
 
-import step3.model.Car
-
 object RandomEngineCarFactory : CarFactory {
     override fun create() = Car(RandomEngine)
 }
