+++ src/test/kotlin/step3/domain/CarFactoryTest.kt
@@ -2,7 +2,6 @@ package step3.domain
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
-import step3.model.Car
 
 class CarFactoryTest {
     @Test
